+++ finality-verifiers/grandpa/src/lib.rs
@@ -247,7 +247,7 @@ pub mod pallet {
     #[pallet::storage]
     #[pallet::getter(fn get_imported_headers)]
     pub(super) type ImportedHeaders<T: Config<I>, I: 'static = ()> =
-        StorageMap<_, Blake2_256, BridgedBlockHash<T, I>, BridgedHeader<T, I>>;
+        StorageMap<_, Identity, BridgedBlockHash<T, I>, BridgedHeader<T, I>>;
 
     #[pallet::storage]
     pub(super) type RelayChainId<T: Config<I>, I: 'static = ()> =
@@ -261,7 +261,7 @@ pub mod pallet {
     /// Maps a parachain chain_id to the corresponding chain ID.
     #[pallet::storage]
     pub(super) type ParachainIdMap<T: Config<I>, I: 'static = ()> =
-        StorageMap<_, Blake2_256, ChainId, ParachainRegistrationData>;
+        StorageMap<_, Identity, ChainId, ParachainRegistrationData>;
 
     /// Optional pallet owner.
     ///
@@ -343,6 +343,8 @@ pub mod pallet {
             ensure_root(origin)?;
             <EverInitialized<T, I>>::kill();
             <BestFinalizedHash<T, I>>::kill();
+            <ParachainIdMap<T, I>>::drain();
+            <ImportedHeaders<T, I>>::drain();
             <InitialHash<T, I>>::kill();
             <ImportedHashesPointer<T, I>>::kill(); // one ahead of first value
             <RelayChainId<T, I>>::kill();
@@ -1185,6 +1187,30 @@ pub mod tests {
         run_test(|| {
             assert_ok!(initialize_relaychain(Origin::root()));
             assert_ok!(initialize_parachain(Origin::root()));
+        })
+    }
+    use hex_literal::hex;
+    #[test]
+    fn can_register_again_after_reset_with_valid_data_and_signer() {
+        run_test(|| {
+            assert_ok!(initialize_relaychain(Origin::root()));
+            assert_ok!(initialize_parachain(Origin::root()));
+            assert_eq!(
+                InitialHash::<TestRuntime>::get(),
+                Some(
+                    hex!("dcdd89927d8a348e00257e1ecc8617f45edb5118efff3ea2f9961b2ad9b7690a").into()
+                )
+            );
+            assert_ok!(Pallet::<TestRuntime>::reset(Origin::root()));
+            assert_eq!(InitialHash::<TestRuntime>::get(), None);
+            assert_ok!(initialize_relaychain(Origin::root()));
+            assert_ok!(initialize_parachain(Origin::root()));
+            assert_eq!(
+                InitialHash::<TestRuntime>::get(),
+                Some(
+                    hex!("dcdd89927d8a348e00257e1ecc8617f45edb5118efff3ea2f9961b2ad9b7690a").into()
+                )
+            );
         })
     }